requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
faiss-cpu==1.8.0.post1
flake8==7.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2
langsmith==0.1.99
marshmallow==3.21.3
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.3.1
openai==1.41.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyGithub==2.3.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-gitlab==4.9.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
unidiff==0.7.5
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
